Move CandidatesOrdering func to commong package
